bug-hunt #509: virtio - full 64-bit feature negotiation for modern transport; drop dead CTRL_VLAN claim
virtio_pci_modern_init_device() read the device's 64-bit feature word
but negotiated and validated against only the low 32 bits, silently
dropping valid high-half features (VIRTIO_NET_F_RSS bit 60,
VIRTIO_NET_F_HASH_REPORT bit 58) that the header explicitly documents
as negotiable via the modern transport, and falsely failing required
checks that included high bits. Widen supported/required to uint64_t
and negotiate over the full feature space.

virtio_net claimed VIRTIO_NET_F_CTRL_VLAN in VNET_SUPPORTED_FEATURES
but has no VIRTIO_NET_CTRL_VLAN class handling or VLAN filter-table
programming on the control VQ; claiming the feature without the
implementation makes the device drop VLAN-tagged frames. Remove the
dead bit so the device passes all VLAN traffic to the software stack.

+++ src/include/virtio.h
@@ -548,7 +548,7 @@ int virtio_pci_modern_map_bars(struct vpci_modern_device *vdev);
  * 'supported' and 'required' work like virtio_negotiate_features_ex.
  * Returns 0 on success, -1 on failure. */
 int virtio_pci_modern_init_device(struct vpci_modern_device *vdev,
-                                  uint32_t supported, uint32_t required,
+                                  uint64_t supported, uint64_t required,
                                   const struct virtio_feature_entry *feat_table,
                                   const char *driver_name);
